Validar inisiar sesión

+++ eliminar.php
@@ -1,4 +1,11 @@
 <?php
+session_start();
+
+if(!isset($_SESSION['usuario'])){
+    header("location: index.php");
+    exit();
+}
+
 include_once  "conexion.php";
 $conn = getConexion('pokemon');
 $sql = "DELETE FROM pokemones WHERE numero = ?";
